desktop/src/renderer/components/app-project/app-project.spec.ts
import { newSpecPage } from '@stencil/core/testing'
import { AppProject } from './app-project'

describe('app-project', () => {
  describe('normalization', () => {
    it('returns a blank string if the name is undefined', async () => {
      const { rootInstance } = await newSpecPage({
        components: [AppProject],
        html: '<app-project></app-project>',
      })
      expect(rootInstance.normalize(undefined)).toEqual('')
    })

    it('returns a blank string if the name is null', async () => {
      const { rootInstance } = await newSpecPage({
        components: [AppProject],
        html: '<app-project></app-project>',
      })
      expect(rootInstance.normalize(null)).toEqual('')
    })

    it('capitalizes the first letter', async () => {
      const { rootInstance } = await newSpecPage({
        components: [AppProject],
        html: '<app-project></app-project>',
      })
      expect(rootInstance.normalize('quincy')).toEqual('Quincy')
    })

    it('lower-cases the following letters', async () => {
      const { rootInstance } = await newSpecPage({
        components: [AppProject],
        html: '<app-project></app-project>',
      })
      expect(rootInstance.normalize('JOSEPH')).toEqual('Joseph')
    })

    it('handles single letter names', async () => {
      const { rootInstance } = await newSpecPage({
        components: [AppProject],
        html: '<app-project></app-project>',
      })
      expect(rootInstance.normalize('q')).toEqual('Q')
    })
  })
})
